api-chg: Move avc to own package
