+++ avc/avcdecoderconfigurationrecord.go
@@ -1,4 +1,4 @@
-package mp4
+package avc
 
 import (
 	"encoding/binary"
